Move detached buffer to libui
Move it into libui early so that new modifications towards it can be
coded in the libui style

This CL only moves the file and updates Android.bp with proper
dependencies. Will re-format the coding style in a follow up CL. The
reason behind this is to keep this CL small enough so that the "git
mv" operation will be considered as an renaming rather than a complete
rewrite.

Note that DetachedBuffer is not exposed to VNDK, so that we won't
need to worry about ABI compatibility. Also, it temporarily introduces
some clang warning exceptions, we should be able to remove them very
soon once pdx to binder refactor is done for detached buffer.

Bug: 112010261
Test: atest BufferHubMetadata_test

+++ libs/ui/BufferHubMetadata.cpp
@@ -19,7 +19,7 @@
 
 #include <cutils/ashmem.h>
 #include <log/log.h>
-#include <private/dvr/buffer_hub_metadata.h>
+#include <ui/BufferHubMetadata.h>
 
 namespace android {
 namespace dvr {
@@ -71,7 +71,7 @@ BufferHubMetadata BufferHubMetadata::Import(pdx::LocalHandle ashmem_handle) {
     return {};
   }
 
-  size_t metadata_size = ashmem_get_size_region(ashmem_handle.Get());
+  size_t metadata_size = static_cast<size_t>(ashmem_get_size_region(ashmem_handle.Get()));
   size_t user_metadata_size = metadata_size - kMetadataHeaderSize;
 
   // Note that here the buffer state is mapped from shared memory as an atomic
